fix: 이전 메세지 받는 로직 수정
