--- src/hooks/chat/useGetMessages.ts
@@ -1,13 +1,0 @@
-import { getChatMessages } from '@/apis/chat';
-import { CHAT_KEYS } from '@/constants/@queryKeys';
-import { useQuery } from '@tanstack/react-query';
-
-const useGetMessages = (roomId: number) => {
-  return useQuery({
-    queryKey: CHAT_KEYS.CHAT_MESSAGES(roomId),
-    queryFn: () => getChatMessages({ roomId }),
-    enabled: !!roomId,
-  });
-};
-
-export default useGetMessages;
